requirements.txt
streamlit>=1.35
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
joblib>=1.3
matplotlib>=3.7
plotly>=5.18
shap>=0.44
lime>=0.2.0.1
xgboost>=1.7
groq>=0.9
fpdf2>=2.7
pytesseract>=0.3.10
Pillow>=10.0
pdf2image>=1.16
